[scanner] fix: refactor Marketplace.tsx to reduce file size
